requirements.txt
websocket-client
mysql-client
configparser